Attempting to build front-end out using react😟
When the response is received, we call the json method on the response object to parse the response body as JSON. This gives us a JavaScript object that we can use to update the state of the component using the setData function.

The PHP code in controller.class.php is responsible for generating the response that is returned by the server. Without seeing the code in that file, it's difficult to say exactly what it's doing. However, based on the fact that we are calling json on the response object in JavaScript, it's likely that the PHP code is generating a JSON response.